pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "artifice"
version = "0.1.0"
description = "Human in the loop agentic AI for interactive coding"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["textual", "tui", "repl", "python", "interactive", "ai", "agentic"]
dependencies = [
    "textual[syntax]>=7.5.0",
    "textual-autocomplete>=4.0.6",
    "any-llm-sdk>=0.1.0",
    "pyperclip>=1.11.0",
    "pyyaml>=6.0",
]

[project.scripts]
artifice = "artifice.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/artifice"]
